Changes score types from string to decimal
Updates the data type of score properties from string to decimal in relevant DTOs.

This change ensures accurate storage and manipulation of score values, preventing potential data type conversion issues and improving data integrity.

Also replaces the "VNHSI" Facility Type with "VRP".

+++ FMS.Domain/Dto/Status/StatusSummaryDto.cs
@@ -96,7 +96,7 @@ namespace FMS.Domain.Dto
         public GapsAssessment GAPSAssessment { get; set; }
 
         [Display(Name = "Cost Estimate")]
-        public double? CostEstimate { get; set; }
+        public decimal? CostEstimate { get; set; }
 
         [Display(Name = "Cost Estimate Date")]
         public DateOnly? CostEstimateDate { get; set; }
